requirements.txt
asgiref==3.5.0
cffi==1.15.0
cryptography==36.0.2
distlib==0.3.4
Django==3.2.13
django-admin-searchable-dropdown==1.2
django-cors-headers==3.11.0
django-extensions==3.1.5
django-filter==21.1
django-taggit==2.1.0
djangorestframework==3.13.1
filelock==3.4.2
mysql-connector==2.2.9
platformdirs==2.4.1
pycparser==2.21
PyJWT==2.3.0
PyMySQL==1.0.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.2.0
tzdata==2022.1
virtualenv==20.13.0
